# apps/server/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AuctionPricing": {
        "additionalProperties": false,
        "description": "Exact observed auction amounts with currency attached to each value.",
        "example": {
          "baseAmount": {
            "amount": "123456789.00",
            "currency": "KRW"
          },
          "plannedAmount": null
        },
        "properties": {
          "baseAmount": {
            "$ref": "#/components/schemas/Money"
          },
          "plannedAmount": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "baseAmount",
          "plannedAmount"
        ],
        "type": "object"
      },
      "AuctionProvenance": {
        "additionalProperties": false,
        "description": "Traceability from a public auction revision to immutable source evidence and normalization.",
        "properties": {
          "contentSha256": {
            "$ref": "#/components/schemas/ContentSha256"
          },
          "normalizedRecordId": {
            "$ref": "#/components/schemas/PositiveBigintText"
          },
          "observationId": {
            "$ref": "#/components/schemas/PositiveBigintText"
          },
          "sourceSystem": {
            "$ref": "#/components/schemas/SourceSystem"
          }
        },
        "required": [
          "sourceSystem",
          "observationId",
          "normalizedRecordId",
          "contentSha256"
        ],
        "type": "object"
      },
      "AuctionSchedule": {
        "additionalProperties": false,
        "description": "UTC lifecycle instants; null means that instant is not yet observed or not supplied.",
        "properties": {
          "announcedAt": {
            "$ref": "#/components/schemas/InstantText"
          },
          "deadlineAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InstantText"
              }
            ],
            "nullable": true
          },
          "openedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InstantText"
              }
            ],
            "nullable": true
          }
        },
        "required": [
          "announcedAt",
          "deadlineAt",
          "openedAt"
        ],
        "type": "object"
      },
      "CanonicalMoneyAmount": {
        "description": "Canonical nonnegative KRW amount with exactly two fractional digits.",
        "maxLength": 19,
        "pattern": "^(?:0|[1-9][0-9]{0,15})\\.[0-9]{2}$",
        "type": "string"
      },
      "ContentSha256": {
        "description": "Lowercase hexadecimal SHA-256 digest of immutable source content.",
        "pattern": "^[0-9a-f]{64}$",
        "type": "string"
      },
      "EatbidApiV1Auction": {
        "additionalProperties": false,
        "properties": {
          "identity": {
            "$ref": "#/components/schemas/PublicAuctionIdentity"
          },
          "pricing": {
            "$ref": "#/components/schemas/AuctionPricing"
          },
          "provenance": {
            "$ref": "#/components/schemas/AuctionProvenance"
          },
          "schedule": {
            "$ref": "#/components/schemas/AuctionSchedule"
          }
        },
        "required": [
          "identity",
          "schedule",
          "pricing",
          "provenance"
        ],
        "type": "object"
      },
      "ExternalBidId": {
        "description": "Opaque source-scoped auction identifier; never parsed for identity.",
        "maxLength": 512,
        "minLength": 1,
        "type": "string"
      },
      "InstantText": {
        "description": "Canonical UTC ISO 8601 instant text ending in Z with at most nanosecond precision.",
        "example": "2026-08-30T00:00:00Z",
        "maxLength": 35,
        "pattern": "^(?:(?:[0-9][0-9][2468][048]|[0-9][0-9][13579][26]|[0-9][0-9]0[48]|[02468][048]00|[13579][26]00)-02-29|[0-9]{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12][0-9]|3[01])|(?:0[469]|11)-(?:0[1-9]|[12][0-9]|30)|(?:02)-(?:0[1-9]|1[0-9]|2[0-8])))T(?:(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](?:\\.[0-9]{0,8}[1-9])?(?:Z))$",
        "type": "string"
      },
      "LiveHealth": {
        "additionalProperties": false,
        "description": "The server process can handle HTTP work.",
        "properties": {
          "status": {
            "enum": [
              "live"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "Money": {
        "additionalProperties": false,
        "description": "Exact KRW money value; amount remains decimal text across JSON boundaries.",
        "properties": {
          "amount": {
            "$ref": "#/components/schemas/CanonicalMoneyAmount"
          },
          "currency": {
            "enum": [
              "KRW"
            ],
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency"
        ],
        "type": "object"
      },
      "PositiveBigintText": {
        "description": "Lossless positive PostgreSQL bigint encoded as canonical decimal text.",
        "maxLength": 19,
        "pattern": "^[1-9][0-9]*$",
        "type": "string"
      },
      "ProblemCode": {
        "enum": [
          "VALIDATION_ERROR",
          "UNAUTHENTICATED",
          "FORBIDDEN",
          "NOT_FOUND",
          "AUCTION_NOT_FOUND",
          "CONFLICT",
          "RATE_LIMITED",
          "DEPENDENCY_UNAVAILABLE",
          "INTERNAL_ERROR"
        ],
        "type": "string"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "format": "uri",
            "maxLength": 256,
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "PublicAuctionIdentity": {
        "additionalProperties": false,
        "description": "Public identity and display state for one canonical auction revision.",
        "properties": {
          "auctionId": {
            "$ref": "#/components/schemas/PositiveBigintText"
          },
          "displayBidNumber": {
            "maxLength": 128,
            "minLength": 1,
            "nullable": true,
            "type": "string"
          },
          "externalBidId": {
            "$ref": "#/components/schemas/ExternalBidId"
          },
          "revisionId": {
            "$ref": "#/components/schemas/PositiveBigintText"
          },
          "status": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "auctionId",
          "revisionId",
          "externalBidId",
          "displayBidNumber",
          "title",
          "status"
        ],
        "type": "object"
      },
      "ReadyHealth": {
        "additionalProperties": false,
        "description": "The server is accepting application work.",
        "properties": {
          "status": {
            "enum": [
              "ready"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "SourceSystem": {
        "description": "Bounded identifier of the system that supplied an observation.",
        "maxLength": 64,
        "minLength": 1,
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Bounded canonical HTTP contracts for the eatbid server.",
    "title": "eatbid API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/auctions/{auctionId}": {
      "get": {
        "operationId": "findAuction",
        "parameters": [
          {
            "example": "9007199254740993",
            "in": "path",
            "name": "auctionId",
            "required": true,
            "schema": {
              "pattern": "^[1-9][0-9]*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EatbidApiV1Auction"
                }
              }
            },
            "description": "Canonical auction"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid auction ID"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Auction not found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server defect"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "Find a canonical auction",
        "tags": [
          "procurement"
        ]
      }
    },
    "/health/live": {
      "get": {
        "operationId": "healthLive",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveHealth"
                }
              }
            },
            "description": "Process is live"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server defect"
          }
        },
        "summary": "Process liveness",
        "tags": [
          "operations"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "healthReady",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyHealth"
                }
              }
            },
            "description": "Application is ready"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server defect"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Application dependency is unavailable"
          }
        },
        "summary": "Application readiness",
        "tags": [
          "operations"
        ]
      }
    }
  }
}